Move logic into single place.
