+++ std-bits/table/src/main/java/org/enso/table/data/index/UnorderedMultiValueKey.java
@@ -1,8 +1,5 @@
 package org.enso.table.data.index;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -39,7 +36,7 @@ public class UnorderedMultiValueKey extends MultiValueKeyBase {
       Object value = this.get(i);
       if (value != null) {
         hasFloatValues = hasFloatValues || NumericConverter.isDecimalLike(value);
-        Object folded = foldObject(value, textFoldingStrategy.get(i));
+        Object folded = EnsoObjectWrapper.foldObject(value, textFoldingStrategy.get(i));
         h += folded.hashCode();
       }
     }
@@ -49,69 +46,7 @@ public class UnorderedMultiValueKey extends MultiValueKeyBase {
   }
 
   protected Object getObjectFolded(int index) {
-    return foldObject(this.get(index), textFoldingStrategy.get(index));
-  }
-
-  /**
-   * Folds the value to ensure consistency with Enso's equality.
-   *
-   * <p>Case-sensitivity of text folding is controlled by {@code textFoldingStrategy}.
-   */
-  protected static Object foldObject(Object value, TextFoldingStrategy textFoldingStrategy) {
-    if (value == null) {
-      return null;
-    }
-
-    if (value instanceof String s) {
-      return textFoldingStrategy.fold(s);
-    }
-
-    Object numeric = foldNumeric(value);
-    if (numeric != null) {
-      return numeric;
-    }
-
-    if (value instanceof Boolean) {
-      return value;
-    }
-
-    if (value instanceof LocalDate
-        || value instanceof LocalTime
-        || value instanceof ZonedDateTime) {
-      return value;
-    }
-
-    return new EnsoObjectWrapper(value);
-  }
-
-  /**
-   * If the value is a numeric type, this method coerces it in such a way to ensure consistency with
-   * Enso.
-   *
-   * <p>Integer types are coerced to {@code Long} and floating point values are coerced to {@code
-   * Double} unless they represent a whole integer in which case they are also coerced to {@code
-   * Long}, to ensure the Enso property that {@code 2 == 2.0}.
-   *
-   * <p>Returns {@code null} if the value was not a numeric value.
-   */
-  protected static Object foldNumeric(Object value) {
-    if (value instanceof Long) {
-      return value;
-    } else if (value instanceof Integer i) {
-      return i.longValue();
-    } else if (value instanceof Byte b) {
-      return b.longValue();
-    } else if (value instanceof Float f && f % 1 == 0) {
-      return f.longValue();
-    } else if (value instanceof Double d && d % 1 == 0) {
-      return d.longValue();
-    } else if (value instanceof Float f) {
-      return f.doubleValue();
-    } else if (value instanceof Double d) {
-      return d;
-    }
-
-    return null;
+    return EnsoObjectWrapper.foldObject(this.get(index), textFoldingStrategy.get(index));
   }
 
   @Override
