+++ std-bits/table/src/main/java/org/enso/table/data/table/join/scan/MatcherFactory.java
@@ -6,7 +6,6 @@ import java.util.stream.Collectors;
 
 import org.enso.base.ObjectComparator;
 import org.enso.base.Text_Utils;
-import org.enso.base.polyglot.EnsoObjectWrapper;
 import org.enso.base.polyglot.NumericConverter;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.StringStorage;
@@ -87,9 +86,7 @@ public class MatcherFactory {
         problems.add(new FloatingPointGrouping(rightColumnName, right));
       }
 
-      // We could do a fast-path for some known primitive types, but it doesn't matter as it will be
-      // replaced with hashing soon anyway.
-      return new EnsoObjectWrapper(leftValue).equals(new EnsoObjectWrapper(rightValue));
+      return ObjectComparator.areEqual(leftValue, rightValue);
     }
 
     @Override
@@ -160,8 +157,6 @@ public class MatcherFactory {
         return false;
       }
 
-      // We could do a fast-path for some known primitive types, but it doesn't matter as it should
-      // be replaced with sorting optimization soon(ish).
       return ObjectComparator.DEFAULT.compare(leftValue, rightLowerValue) >= 0
           && ObjectComparator.DEFAULT.compare(leftValue, rightUpperValue) <= 0;
     }
